poncho/poncho/requirements.txt
asgiref==3.7.2
Django==4.0.2
django-environ==0.11.2
django-prometheus==2.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
prometheus-client==0.19.0
pymongo==3.12.1
python-dotenv==1.0.1
pytz==2021.3
sqlparse==0.2.4
typing_extensions==4.9.0
